+++ src/zkevm_specs/__init__.py
@@ -1,10 +1,9 @@
-from . import bytecode
+from . import bytecode_circuit
 from . import copy_circuit
 from . import encoding
 from . import evm
 from . import exp_circuit
 from . import opcode
 from . import state_circuit
-from . import tx
+from . import tx_circuit
 from . import util
-from . import tx
