+++ tests/test_bytecode_circuit.py
@@ -1,7 +1,7 @@
 import traceback
 from copy import deepcopy
 
-from zkevm_specs.bytecode import *
+from zkevm_specs.bytecode_circuit import *
 from zkevm_specs.evm import Opcode, Bytecode, BytecodeFieldTag, BytecodeTableRow, is_push
 from zkevm_specs.util import RLC, rand_fq, U256
